backend/requirements.txt
fastapi==0.115.2
uvicorn==0.32.0
sqlalchemy==2.0.36
pgvector
psycopg2-binary==2.9.10
alembic==1.13.3
pydantic-settings==2.5.2
httpx==0.27.2
python-jose==3.3.0
tenacity==9.0.0
google-auth
google-auth-oauthlib